mcp-client-agent-llamaindex/pyproject.toml
[project]
name = "mcp-client-agent-llamaindex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncio>=3.4.3",
    "llama-index-core>=0.12.43",
    "llama-index-llms-ollama>=0.6.2",
    "llama-index-tools-mcp>=0.2.5",
    "mcp>=1.9.4",
    "ollama>=0.5.1",
    "python-dotenv>=1.1.0",
]
